# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SistemaJava • Mercadinho PDV",
    "description": "Obtenha CSRF em /api/v1/auth/csrf; envie X-XSRF-TOKEN em mutações. Login retorna Bearer JWT.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/customers/{id}": {
      "put": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "products",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/customers": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCustomer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{id}/bottles": {
      "get": {
        "tags": [
          "extras-controller"
        ],
        "operationId": "customerBottles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "extras-controller"
        ],
        "operationId": "bottle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BottleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/payments": {
      "post": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "pay_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditPayment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "categories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/analytics": {
      "get": {
        "tags": [
          "extras-controller"
        ],
        "operationId": "analytics",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/finance/statements": {
      "get": {
        "tags": [
          "finance-controller"
        ],
        "operationId": "statement",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/customers/{id}/titles": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "titles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "statement_1",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Credit"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductInput": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "barcode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "unit": {
            "type": "string",
            "pattern": "UN|KG"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "cost": {
            "type": "number",
            "minimum": 0
          },
          "minimumStock": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "cost",
          "minimumStock",
          "name",
          "price",
          "unit"
        ]
      },
      "ProductView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryName": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "minimumStock": {
            "type": "number"
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "CustomerInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "document": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "minLength": 0
          },
          "creditLimit": {
            "type": "number",
            "minimum": 0
          },
          "address": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "termDays": {
            "type": "integer",
            "format": "int32",
            "maximum": 365,
            "minimum": 0
          },
          "interestDay": {
            "type": "number",
            "maximum": 10,
            "minimum": 0
          },
          "penaltyDay": {
            "type": "number",
            "maximum": 10,
            "minimum": 0
          }
        },
        "required": [
          "creditLimit",
          "name"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "creditLimit": {
            "type": "number"
          },
          "balance": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "termDays": {
            "type": "integer",
            "format": "int32"
          },
          "interestDay": {
            "type": "number"
          },
          "penaltyDay": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "CategoryInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "markup": {
            "type": "number",
            "maximum": 1000,
            "minimum": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "markup": {
            "type": "number"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "Sale": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestHash": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "total": {
            "type": "number"
          },
          "fees": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "BottleInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 40,
            "minLength": 0
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "maximum": 1000,
            "minimum": 1
          },
          "direction": {
            "type": "string",
            "pattern": "TAKEN|RETURNED"
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "direction",
          "quantity"
        ]
      },
      "CreditPayment": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "method": {
            "type": "string",
            "pattern": "CASH|PIX|DEBIT|CREDIT"
          },
          "waiveCharges": {
            "type": "boolean"
          }
        },
        "required": [
          "amount",
          "customerId",
          "description",
          "requestId"
        ]
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageProductView": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductView"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageCustomer": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "Credit": {
        "type": "object",
        "properties": {
          "paymentRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "sale": {
            "$ref": "#/components/schemas/Sale"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "amount": {
            "type": "number"
          },
          "remaining": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "interestDay": {
            "type": "number"
          },
          "penaltyDay": {
            "type": "number"
          },
          "method": {
            "type": "string"
          },
          "charges": {
            "type": "number"
          },
          "cashSessionId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}